Change Histogram return type to float64

+++ database/in_memory_database_query.go
@@ -193,7 +193,7 @@ func (s *inMemoryDatabaseQuery) GroupAggregation(field, function string, keys ..
 }
 
 // Histogram returns a time series data points based on the time field, supported intervals: Minute, Hour, Day, week, month
-func (s *inMemoryDatabaseQuery) Histogram(field, function, timeField string, interval time.Duration, keys ...string) (out map[Timestamp]int64, total int64, err error) {
+func (s *inMemoryDatabaseQuery) Histogram(field, function, timeField string, interval time.Duration, keys ...string) (out map[Timestamp]float64, total float64, err error) {
 	return nil, 0, fmt.Errorf("not yet implemented")
 }
